interrupt: move pre-NMI functions to interrupt.h
pre-NMI functions were added to exception.h but it was a mistake:
all other interrupts are handled in interrupt.h, so it makes sense
to handle pre-NMI there as well.
